requirements.txt
numba
tokenizers
tqdm
requests
boto3
filelock
ruamel.yaml
wget
pathos
tree-sitter==0.2.2
jsonlines
dpu_utils
sentencepiece
dgl-cu102
rouge
nltk
jsbeautifier
loguru
gpustat
cython
gdown
Deprecated
h5py
tables
mkdocs